pdf_report_generator/requirements.txt
reportlab>=4.0
matplotlib>=3.7
Pillow>=10.0
pypdf>=3.0
numpy>=1.24
